requirements.txt
datefinder==0.7.3
Flask==2.2.2
locationtagger==0.0.1
nltk==3.8.1
python-dotenv==1.0.0
Requests==2.31.0
Werkzeug==2.2.2
